Finish the module placement: move SSOViewsController + RequestHelpers into Http, dissolve the Kernel test folder [#867]
The #790/#807 module migration left two stragglers and a legacy test folder:

- SSOViewsController lived in SSO-Auth/Views/ under namespace ...Views; it is a
  web-boundary controller and belongs in the Http module. Moved to
  SSO-Auth/Api/Http/ with namespace ...Api.Http. It imports no Api module, so it
  adds no edge to the module DAG. Its route is [Route("[controller]")] —
  derived from the class name, not the namespace — so the served URL /SSOViews
  is unchanged (SSOViewsControllerTests exercise GetView and pass).
- RequestHelpers already sat physically in Api/Http/ but kept the stale
  namespace ...Helpers (its sibling ProviderConnectionTester was already
  ...Api.Http). Corrected to ...Api.Http; the now-redundant `using ...Helpers`
  is dropped from SSOController (same namespace) and the two test consumers
  updated. This drift was invisible to the conformance suite.
- Dissolved SSO-Auth.Tests/Kernel/: the SSOController* split tests and
  SSOViewsControllerTests move to SSO-Auth.Tests/Http/ (next to the source they
  cover), and SSOPluginPageManifestTests to the test root (it covers the root
  SSOPlugin type). The test namespace is flat, so these are pure git mv with no
  code change. ModuleTests_MirrorTheSourceModuleFolders (#791) now requires and
  gets SSOViewsControllerTests under Http/; its comment is updated to drop the
  stale Kernel reference.

No behaviour change: all 1786 tests pass on net9.0 and net10.0, including the
architecture-conformance suite (module DAG, flat-Api-empty, folder mirror).

Closes #867
