+++ SSO-Auth.Tests/Http/SSOControllerScopeStringTests.cs
