+++ SSO-Auth.Tests/SSOPluginPageManifestTests.cs
